@@ -2,9 +2,10 @@ package com.example.mylauncher;
 
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentStatePagerAdapter;
+
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentStatePagerAdapter;
 
 public class HomeScreenPageAdapter extends FragmentStatePagerAdapter {
     private int maxscreens = 2;
