@@ -1,9 +1,13 @@
 package com.example.mylauncher.utils;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
 import android.util.AttributeSet;
 import android.view.View;
+import android.view.ViewGroup;
+import android.view.animation.GridLayoutAnimationController;
+
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 public class RecyclerViewEmptySupport extends RecyclerView {
     private View emptyView;
@@ -50,5 +54,51 @@ public class RecyclerViewEmptySupport extends RecyclerView {
 
     public void setEmptyView(View emptyView) {
         this.emptyView = emptyView;
+    }
+
+    /**
+     * Based on:
+     * https://gist.github.com/Musenkishi/8df1ab549857756098ba
+     * Credit to Freddie (Musenkishi) Lust-Hed
+     * <p>
+     * ...which in turn is based on the GridView implementation of attachLayoutParameters(...):
+     * https://android.googlesource.com/platform/frameworks/base/+/master/core/java/android/widget/GridView.java
+     */
+    @Override
+    protected void attachLayoutAnimationParameters(View child, ViewGroup.LayoutParams params,
+                                                   int index, int count) {
+        final LayoutManager layoutManager = getLayoutManager();
+        if (getAdapter() != null && layoutManager instanceof GridLayoutManager) {
+
+            GridLayoutAnimationController.AnimationParameters animationParams =
+                    (GridLayoutAnimationController.AnimationParameters) params.layoutAnimationParameters;
+
+            if (animationParams == null) {
+                // If there are no animation parameters, create new once and attach them to
+                // the LayoutParams.
+                animationParams = new GridLayoutAnimationController.AnimationParameters();
+                params.layoutAnimationParameters = animationParams;
+            }
+
+            // Next we are updating the parameters
+
+            // Set the number of items in the RecyclerView and the index of this item
+            animationParams.count = count;
+            animationParams.index = index;
+
+            // Calculate the number of columns and rows in the grid
+            final int columns = ((GridLayoutManager) layoutManager).getSpanCount();
+            animationParams.columnsCount = columns;
+            animationParams.rowsCount = count / columns;
+
+            // Calculate the column/row position in the grid
+            final int invertedIndex = count - 1 - index;
+            animationParams.column = columns - 1 - (invertedIndex % columns);
+            animationParams.row = animationParams.rowsCount - 1 - invertedIndex / columns;
+
+        } else {
+            // Proceed as normal if using another type of LayoutManager
+            super.attachLayoutAnimationParameters(child, params, index, count);
+        }
     }
 }
